+++ src/modules/text-input/text-input.tsx
@@ -24,6 +24,10 @@ const StyledInput = Styled.TextInput<IStyledInput>`
   width: 100%;
   border-radius: 3px;
 
+  font-family: ${({theme, $font = 'primary'}) => {
+    return theme.typography.faces[`${$font}Regular`];
+  }};
+
   font-size: ${({$isUnderlined}) => ($isUnderlined ? '20px' : '16px')};
 
   padding-bottom: ${({multiline}) => (multiline ? majorScale(1, 'px') : 0)};
@@ -58,6 +62,7 @@ const StyledInput = Styled.TextInput<IStyledInput>`
   border-top-width: ${({$isUnderlined}) => ($isUnderlined ? 0 : '1px')};
   border-right-width: ${({$isUnderlined}) => ($isUnderlined ? 0 : '1px')};
   border-left-width: ${({$isUnderlined}) => ($isUnderlined ? 0 : '1px')};
+  text-align-vertical: ${({multiline}) => (multiline ? 'top' : 'center')};
 `;
 
 // Component
@@ -65,6 +70,7 @@ export const TextInput: React.FC<IProps> = ({
   label,
   required,
   disabled,
+  font = 'primary',
   placeholder,
   isUnderlined,
   onBlur,
@@ -123,6 +129,7 @@ export const TextInput: React.FC<IProps> = ({
 
       <StyledInput
         testID="input"
+        $font={font}
         $isFocused={isFocused}
         $isUnderlined={isUnderlined}
         editable={!disabled}