+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### v0.0.42
 
 - Update Person Record Component, Fixes UI and Functionalities.
+- Add SelectPicker showCustomIcon Prop
+- Add style textAlignVertical:'top' to TextInput when multiline for same behavior in both platforms
 
 --
 
@@ -10,6 +12,7 @@
 
 - Add Note list item component
 - Add SelectPicker Required Prop
+- Add fontFamily on text Input
 
 --